Cadastro de plano

+++ companhiaTelefonica/src/companhiaTelefonica/Plano.java
@@ -4,11 +4,11 @@ public class Plano extends companhiaEletronica
 {
 	private int codigo;
 	private String descricao;
-	private int valorMensal;
-	private int franquia;
-	private int valorChamada;
+	private float valorMensal;
+	private float franquia;
+	private float valorChamada;
 
-	public Plano(int codigo ,String descricao, int valorMensal,int franquia,int valorChamada)
+	public Plano(int codigo ,String descricao, float valorMensal,float franquia,float valorChamada)
 	{
 		this.codigo=codigo;
 		this.descricao  = descricao;
